+++ src/clj/rems/api/services/todos.clj
@@ -54,7 +54,8 @@
           :application.command/remove-member
           :application.command/uninvite-member
           ;; remarks can be made without a request, also on handled todos
-          :application.command/remark}
+          :application.command/remark
+          :application.command/send-expiration-notifications}
         all-commands (set (keys commands/command-schemas))]
 
     ;; This test is to make sure that as new commands are added,