+++ test/clj/rems/api/test_pdf.clj
@@ -26,7 +26,8 @@
         res-id1 (test-helpers/create-resource! nil)
         item-id1 (test-helpers/create-catalogue-item! {:form-id form :workflow-id wfid :resource-id res-id1})
         app-id (test-helpers/create-draft! "applicant" [item-id1] "draft")]
-    (test-helpers/submit-application app-id "applicant")))
+    (test-helpers/submit-application {:application-id app-id
+                                      :actor "applicant"})))
 
 (deftest test-experimental-pdf-smoke
   ;; need to spin up an actual http server so that something can serve